+++ Intellishelf.Domain/Books/Models/PagedResult.cs
@@ -1,9 +1,9 @@
 namespace Intellishelf.Domain.Books.Models;
 
-public class PagedResult<T>(IReadOnlyCollection<T> items, int totalCount, int page, int pageSize)
+public class PagedResult<T>(IReadOnlyCollection<T> items, long totalCount, int page, int pageSize)
 {
     public IReadOnlyCollection<T> Items { get; } = items;
-    public int TotalCount { get; } = totalCount;
+    public long TotalCount { get; } = totalCount;
     public int Page { get; } = page;
     public int PageSize { get; } = pageSize;
     public int TotalPages { get; } = (int)Math.Ceiling(totalCount / (double)pageSize);
